fix: refactor GET and PUT api code
