+++ cms/djangoapps/contentstore/rest_api/v1/serializers/__init__.py
@@ -5,7 +5,7 @@ from .certificates import CourseCertificatesSerializer
 from .course_details import CourseDetailsSerializer
 from .course_index import CourseIndexSerializer
 from .course_rerun import CourseRerunSerializer
-from .course_team import CourseTeamSerializer, CourseTeamManagementSerializer
+from .course_team import CourseTeamSerializer
 from .course_waffle_flags import CourseWaffleFlagsSerializer
 from .grading import CourseGradingModelSerializer, CourseGradingSerializer
 from .group_configurations import CourseGroupConfigurationsSerializer
