+++ cms/djangoapps/contentstore/rest_api/v1/serializers/course_team.py
@@ -3,7 +3,6 @@ API Serializers for course team
 """
 
 from rest_framework import serializers
-from openedx.core.djangoapps.content.course_overviews.models import CourseOverview
 
 
 class UserCourseTeamSerializer(serializers.Serializer):
@@ -19,24 +18,3 @@ class CourseTeamSerializer(serializers.Serializer):
     show_transfer_ownership_hint = serializers.BooleanField()
     users = UserCourseTeamSerializer(many=True)
     allow_actions = serializers.BooleanField()
-
-
-class CourseTeamManagementSerializer(serializers.ModelSerializer):
-    """Serializer for course team management context data"""
-    role = serializers.SerializerMethodField()
-
-    class Meta:
-        model = CourseOverview
-        fields = ("id", "display_name", "role")
-
-    def get_role(self, obj):
-        course_role_map = self.context.get("course_role_map", {})
-        return course_role_map.get(str(obj.id))
-
-    def to_representation(self, instance):
-        data = super().to_representation(instance)
-        return {
-            "course_id": data["id"],
-            "course_name": data["display_name"],
-            "role": data["role"],
-        }
