+++ lms/djangoapps/support/rest_api/__init__.py
@@ -1,0 +1,1 @@
+
