+++ lms/djangoapps/support/rest_api/v1/__init__.py
@@ -1,0 +1,1 @@
+
