+++ lms/djangoapps/support/tests/test_views.py
@@ -45,11 +45,14 @@ from common.djangoapps.student.models import (
     CourseEnrollmentAttribute,
     ManualEnrollmentAudit
 )
-from common.djangoapps.student.roles import GlobalStaff, SupportStaffRole
+from common.djangoapps.student.roles import  (
+    GlobalStaff,
+    SupportStaffRole
+)
 from common.djangoapps.student.tests.factories import (
-    CourseEnrollmentFactory,
     CourseEnrollmentAttributeFactory,
-    UserFactory,
+    CourseEnrollmentFactory,
+    UserFactory
 )
 from common.djangoapps.third_party_auth.tests.factories import SAMLProviderConfigFactory
 from common.test.utils import disable_signal
